Update some of the integer types we use internally to remove the "implicit conversion shortens 64-bit value into a 32-bit value" warnings.  Many of these were due to using integers to store values better characterized as size_t's or other bitness-agnostic types.
Yay for 64-bit operating systems!

* subversion/libsvn_subr/xml.c,
  subversion/libsvn_subr/macos_keychain.c,
  subversion/libsvn_ra_local/ra_plugin.c,
  subversion/libsvn_ra_svn/cyrus_auth.c,
  subversion/libsvn_fs_base/bdb/dbt.c,
  subversion/libsvn_fs_base/bdb/dbt.h,
  subversion/libsvn_fs_base/bdb/uuids-table.c,
  subversion/libsvn_fs_base/bdb/rev-table.c,
  subversion/libsvn_fs_util/fs-util.c,
  subversion/libsvn_repos/load.c,
  subversion/libsvn_repos/log.c,
  subversion/libsvn_repos/repos.c,
  subversion/libsvn_repos/commit.c,
  subversion/libsvn_repos/replay.c,
  subversion/libsvn_delta/svndiff.c,
  subversion/libsvn_fs_fs/fs_fs.c,
  subversion/libsvn_fs_fs/dag.c:
    Update various types to be 64 bits wide, where appropriate.  Where not
    possible (due to external or API requirements) cast.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@879455 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_fs_util/fs-util.c
@@ -38,7 +38,7 @@ const char *
 svn_fs__canonicalize_abspath(const char *path, apr_pool_t *pool)
 {
   char *newpath;
-  int path_len;
+  size_t path_len;
   int path_i = 0, newpath_i = 0;
   svn_boolean_t eating_slashes = FALSE;
